refactor: split into smaller components for postcard
