+++ src/components/auth/SecureRoute.tsx
