+++ src/components/common/Modal.tsx
@@ -42,7 +42,7 @@ const Modal: React.FC<Props> = ({
             {/* Top Right Close (X) Button */}
             <button
               onClick={closeModal}
-              className='absolute  top-4 right-4 text-xl text-gray-500 hover:text-gray-700'
+              className='absolute top-4 right-4 text-xl text-gray-500 hover:text-gray-700'
             >
               <X className='h-6 w-6 cursor-pointer' strokeWidth={2} />
             </button>
