+++ src/components/imageSlider.tsx
