+++ src/pages/example/Example.tsx
@@ -31,7 +31,7 @@ const samplePosts = [
       'https://images.unsplash.com/photo-1518837695005-2083093ee35b?w=400',
       'https://images.unsplash.com/photo-1519904981063-b0cf448d479e?w=400',
     ],
-    disasterType: 'flood' as const,
+    disasterType: 'storm' as const,
     upvotes: 3800,
     downvotes: 1200,
     comments: 8120,
@@ -53,7 +53,7 @@ const samplePosts = [
       'https://images.unsplash.com/photo-1547036967-23d11aacaee0?w=400',
       'https://images.unsplash.com/photo-1518837695005-2083093ee35b?w=400',
     ],
-    disasterType: 'earthquake' as const,
+    disasterType: 'flood' as const,
     upvotes: 1000,
     downvotes: 1200,
     comments: 4350,
